requirements.txt
# Dev environment
pip
autopep8

# App
Flask
gunicorn
twilio
resend
google-cloud-secret-manager
google-cloud-datastore==2.15.0